requirements.txt
altgraph==0.17
cycler==0.10.0
future==0.18.2
httplib2==0.18.1
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
pandas==1.0.3
pefile==2019.4.18
Pillow==7.1.2
plantuml==0.3.0
PyInstaller==3.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pywin32-ctypes==0.2.0
six==1.15.0
